+++ src/types.ts
@@ -2496,6 +2496,8 @@ export type Query = {
   metadataStandard?: Maybe<MetadataStandard>;
   /** Search for a metadata standard */
   metadataStandards?: Maybe<MetadataStandardSearchResults>;
+  /** return all metadata standards whose unique uri values are provided */
+  metadataStandardsByURIs?: Maybe<Array<MetadataStandard>>;
   /** Get all of the user's projects */
   myProjects?: Maybe<ProjectSearchResults>;
   /** Get the Templates that belong to the current user's affiliation (user must be an Admin) */
@@ -2558,6 +2560,8 @@ export type Query = {
   relatedWorksByProject?: Maybe<RelatedWorkSearchResults>;
   /** Search for a repository */
   repositories?: Maybe<RepositorySearchResults>;
+  /** return all repositories whose unique uri values are provided */
+  repositoriesByURIs?: Maybe<Array<Repository>>;
   /** Fetch a specific repository */
   repository?: Maybe<Repository>;
   /** return all distinct subject area keywords across all repositories */
@@ -2706,6 +2710,11 @@ export type QueryMetadataStandardsArgs = {
   paginationOptions?: InputMaybe<PaginationOptions>;
   researchDomainId?: InputMaybe<Scalars['Int']['input']>;
   term?: InputMaybe<Scalars['String']['input']>;
+};
+
+
+export type QueryMetadataStandardsByUrIsArgs = {
+  uris: Array<Scalars['String']['input']>;
 };
 
 
@@ -2863,6 +2872,11 @@ export type QueryRelatedWorksByProjectArgs = {
 
 export type QueryRepositoriesArgs = {
   input: RepositorySearchInput;
+};
+
+
+export type QueryRepositoriesByUrIsArgs = {
+  uris: Array<Scalars['String']['input']>;
 };
 
 
@@ -5872,6 +5886,7 @@ export type QueryResolvers<ContextType = MyContext, ParentType extends Resolvers
   memberRoles?: Resolver<Maybe<Array<Maybe<ResolversTypes['MemberRole']>>>, ParentType, ContextType>;
   metadataStandard?: Resolver<Maybe<ResolversTypes['MetadataStandard']>, ParentType, ContextType, RequireFields<QueryMetadataStandardArgs, 'uri'>>;
   metadataStandards?: Resolver<Maybe<ResolversTypes['MetadataStandardSearchResults']>, ParentType, ContextType, Partial<QueryMetadataStandardsArgs>>;
+  metadataStandardsByURIs?: Resolver<Maybe<Array<ResolversTypes['MetadataStandard']>>, ParentType, ContextType, RequireFields<QueryMetadataStandardsByUrIsArgs, 'uris'>>;
   myProjects?: Resolver<Maybe<ResolversTypes['ProjectSearchResults']>, ParentType, ContextType, Partial<QueryMyProjectsArgs>>;
   myTemplates?: Resolver<Maybe<ResolversTypes['TemplateSearchResults']>, ParentType, ContextType, Partial<QueryMyTemplatesArgs>>;
   myVersionedTemplates?: Resolver<Maybe<Array<Maybe<ResolversTypes['VersionedTemplateSearchResult']>>>, ParentType, ContextType>;
@@ -5903,6 +5918,7 @@ export type QueryResolvers<ContextType = MyContext, ParentType extends Resolvers
   relatedWorksByPlan?: Resolver<Maybe<ResolversTypes['RelatedWorkSearchResults']>, ParentType, ContextType, RequireFields<QueryRelatedWorksByPlanArgs, 'planId'>>;
   relatedWorksByProject?: Resolver<Maybe<ResolversTypes['RelatedWorkSearchResults']>, ParentType, ContextType, RequireFields<QueryRelatedWorksByProjectArgs, 'projectId'>>;
   repositories?: Resolver<Maybe<ResolversTypes['RepositorySearchResults']>, ParentType, ContextType, RequireFields<QueryRepositoriesArgs, 'input'>>;
+  repositoriesByURIs?: Resolver<Maybe<Array<ResolversTypes['Repository']>>, ParentType, ContextType, RequireFields<QueryRepositoriesByUrIsArgs, 'uris'>>;
   repository?: Resolver<Maybe<ResolversTypes['Repository']>, ParentType, ContextType, RequireFields<QueryRepositoryArgs, 'uri'>>;
   repositorySubjectAreas?: Resolver<Maybe<Array<ResolversTypes['String']>>, ParentType, ContextType>;
   researchOutputType?: Resolver<Maybe<ResolversTypes['ResearchOutputType']>, ParentType, ContextType, RequireFields<QueryResearchOutputTypeArgs, 'id'>>;